*Server - 게임 시작 시 이전 상태가 남아있던 버그 수정 - 바나나 생존시간 2초로 변경(다음판에 바나나가 릴리즈 되는 버그 막기 위해서)

+++ Minsik/AnimalKingdom_Server/AnimalKingdom_server/Contents/Player.cpp
@@ -273,7 +273,7 @@ void Player::powerUp()
 		defaultSpeed = MONKEY_SPEED;
 		break;
 	}
-	damage_ = defaultDamage * 1.5;
+	damage_ = int(defaultDamage * 1.5);
 	speed_ = defaultSpeed * 0.5f;
 }
